[Task 3] refactoring
