+++ product-service/src/functions/getProductsById/handler.ts
@@ -3,7 +3,7 @@ import { formatErrorResponse, formatJSONResponse } from '@libs/api-gateway';
 import { middyfy } from '@libs/lambda';
 import { getProductById } from '../../services/products';
 
-export const getProductByIdHandler = middyfy(async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
+export const main = middyfy(async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   try {
     const productId = event.pathParameters.id;
     const product = getProductById(productId);
