+++ product-service/src/functions/getProductsById/index.ts
@@ -1,0 +1,14 @@
+import { handlerPath } from '@libs/handler-resolver';
+
+export const getProductByIdFunction = {
+  handler: `${handlerPath(__dirname)}/handler.main`,
+  events: [
+    {
+      http: {
+        method: 'get',
+        path: 'product/{id}',
+        cors: true,
+      },
+    },
+  ],
+};
