+++ product-service/src/functions/getProductsList/handler.ts
@@ -3,7 +3,7 @@ import { formatJSONResponse } from '@libs/api-gateway';
 import { middyfy } from '@libs/lambda';
 import { getProductsList } from '../../services/products';
 
-export const getProductsListHandler = middyfy(async (): Promise<APIGatewayProxyResult> => {
+export const main = middyfy(async (): Promise<APIGatewayProxyResult> => {
   const productsList = getProductsList();
   return formatJSONResponse(productsList);
 });
