+++ product-service/src/functions/getProductsList/index.ts
@@ -1,0 +1,14 @@
+import { handlerPath } from '@libs/handler-resolver';
+
+export const getProductsFunction = {
+  handler: `${handlerPath(__dirname)}/handler.main`,
+  events: [
+    {
+      http: {
+        method: 'get',
+        path: 'product/available',
+        cors: true,
+      },
+    },
+  ],
+};
