+++ product-service/src/functions/index.ts
@@ -1,0 +1,2 @@
+export { getProductsFunction } from './getProductsList';
+export { getProductByIdFunction } from './getProductsById';
